# swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080/api",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/ticket/inventory" : {
      "put" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "更新票务配置",
        "description" : "更新指定车次和座位类型的票务配置信息，包括总票数、可用票数、票价、状态",
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TicketInventoryCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTicketInventoryDTO"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "创建票务配置",
        "description" : "为指定车次和座位类型创建票务配置，创建后可配置总票数和票价",
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TicketInventoryCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTicketInventoryDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/train/update" : {
      "post" : {
        "tags" : [ "列车管理" ],
        "summary" : "更新列车信息",
        "description" : "修改列车详细信息",
        "operationId" : "updateTrain",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrainInfoDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/train/search" : {
      "post" : {
        "tags" : [ "列车管理" ],
        "summary" : "条件搜索列车",
        "description" : "根据出发站、到达站、日期等条件搜索列车",
        "operationId" : "searchTrains",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrainQueryDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListTrainInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/train/save" : {
      "post" : {
        "tags" : [ "列车管理" ],
        "summary" : "添加列车",
        "description" : "新增列车信息到系统中",
        "operationId" : "addTrain",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrainInfoDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/ticket/inventory/batch" : {
      "post" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "批量创建票务配置",
        "description" : "批量创建多条票务配置，支持一次为多个座位类型创建配置",
        "operationId" : "batchCreate",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/TicketInventoryCreateDTO"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/ticket/add/ticket/{ticketIds}" : {
      "post" : {
        "tags" : [ "票务查询" ],
        "summary" : "手动更新票信息",
        "description" : "手动更新指定车票ID的票信息",
        "operationId" : "updateTicketsByHands",
        "parameters" : [ {
          "name" : "ticketIds",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/train/{id}" : {
      "get" : {
        "tags" : [ "列车管理" ],
        "summary" : "根据ID查询列车",
        "description" : "通过列车ID获取列车详细信息",
        "operationId" : "getTrainById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTrainInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/train/type/{trainType}" : {
      "get" : {
        "tags" : [ "列车管理" ],
        "summary" : "根据列车类型查询",
        "description" : "根据列车类型（高铁、动车、特快等）筛选列车",
        "operationId" : "getTrainsByType",
        "parameters" : [ {
          "name" : "trainType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListTrainInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/train/route" : {
      "get" : {
        "tags" : [ "列车管理" ],
        "summary" : "根据路线查询列车",
        "description" : "根据出发站ID和到达站ID查询经过该路线的所有列车",
        "operationId" : "getTrainsByRoute",
        "parameters" : [ {
          "name" : "startStationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "endStationId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListTrainInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/train/number/{trainNumber}" : {
      "get" : {
        "tags" : [ "列车管理" ],
        "summary" : "根据车次号查询",
        "description" : "通过车次号（如G1234）查询列车信息",
        "operationId" : "getTrainByNumber",
        "parameters" : [ {
          "name" : "trainNumber",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTrainInfoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/ticket/{ticketId}" : {
      "get" : {
        "tags" : [ "票务查询" ],
        "summary" : "获取车票详情",
        "description" : "通过车票ID获取车票详细信息",
        "operationId" : "getTicketDetail",
        "parameters" : [ {
          "name" : "ticketId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTicketInfoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/ticket/inventory/{id}" : {
      "get" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "查询单个票务配置",
        "description" : "根据ID查询票务配置的详细信息",
        "operationId" : "getById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTicketInventoryDTO"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "删除票务配置",
        "description" : "根据ID删除票务配置，删除后该配置将不再生效",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultBoolean"
                }
              }
            }
          }
        }
      }
    },
    "/ticket/inventory/list" : {
      "get" : {
        "tags" : [ "票务配置管理" ],
        "summary" : "条件查询票务配置",
        "description" : "根据车次ID、座位类型等条件查询票务配置列表",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "queryDTO",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/TicketInventoryQueryDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListTicketInventoryDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/train/delete/{id}" : {
      "delete" : {
        "tags" : [ "列车管理" ],
        "summary" : "删除列车",
        "description" : "根据ID删除列车信息",
        "operationId" : "deleteTrain",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ResultString" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TicketInventoryCreateDTO" : {
        "required" : [ "availableCount", "seatType", "totalCount", "trainId" ],
        "type" : "object",
        "properties" : {
          "trainId" : {
            "type" : "integer",
            "description" : "列车ID",
            "format" : "int64",
            "example" : 1
          },
          "seatType" : {
            "type" : "string",
            "description" : "座位类型",
            "example" : "SECOND",
            "enum" : [ "BUSINESS", "FIRST", "SECOND", "SOFT_SLEEPER", "HARD_SLEEPER" ]
          },
          "totalCount" : {
            "minimum" : 0,
            "type" : "integer",
            "description" : "总票数",
            "format" : "int32",
            "example" : 100
          },
          "availableCount" : {
            "minimum" : 0,
            "type" : "integer",
            "description" : "可用票数",
            "format" : "int32",
            "example" : 80
          },
          "price" : {
            "type" : "number",
            "description" : "票价，单位元",
            "example" : 553.5
          },
          "status" : {
            "type" : "integer",
            "description" : "状态 0-正常 1-禁用",
            "format" : "int32",
            "example" : 0
          }
        },
        "description" : "票务配置创建/更新请求"
      },
      "ResultTicketInventoryDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TicketInventoryDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TicketInventoryDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "配置ID",
            "format" : "int64"
          },
          "trainId" : {
            "type" : "integer",
            "description" : "列车ID",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string",
            "description" : "车次号"
          },
          "seatType" : {
            "type" : "string",
            "description" : "座位类型",
            "example" : "SECOND"
          },
          "seatTypeName" : {
            "type" : "string",
            "description" : "座位类型中文名",
            "example" : "二等座"
          },
          "totalCount" : {
            "type" : "integer",
            "description" : "总票数",
            "format" : "int32",
            "example" : 100
          },
          "availableCount" : {
            "type" : "integer",
            "description" : "可用票数",
            "format" : "int32",
            "example" : 80
          },
          "soldCount" : {
            "type" : "integer",
            "description" : "已售票数",
            "format" : "int32",
            "example" : 20
          },
          "price" : {
            "type" : "number",
            "description" : "票价，单位元",
            "example" : 553.5
          },
          "version" : {
            "type" : "integer",
            "description" : "版本号，用于乐观锁",
            "format" : "int32"
          },
          "status" : {
            "type" : "integer",
            "description" : "状态 0-正常 1-禁用",
            "format" : "int32",
            "example" : 0
          },
          "createTime" : {
            "type" : "string",
            "description" : "创建时间",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "description" : "更新时间",
            "format" : "date-time"
          },
          "inventoryPercent" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "description" : "票务配置响应"
      },
...
      "ResultBoolean" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "LocalTime" : {
        "type" : "object",
        "properties" : {
          "hour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minute" : {
            "type" : "integer",
            "format" : "int32"
          },
          "second" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nano" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "TrainInfoDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "trainType" : {
            "type" : "string"
          },
          "startStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "startStationName" : {
            "type" : "string"
          },
          "endStationName" : {
            "type" : "string"
          },
          "startTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "endTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "duration" : {
            "type" : "integer",
            "format" : "int32"
          },
          "durationText" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "stationList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TrainStationDTO"
            }
          },
          "createTime" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updateTime" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "TrainStationDTO" : {
        "type" : "object",
        "properties" : {
          "stationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "stationName" : {
            "type" : "string"
          },
          "stationOrder" : {
            "type" : "integer",
            "format" : "int32"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "departureTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "stopMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isSale" : {
            "type" : "boolean"
          }
        }
      },
      "TrainQueryDTO" : {
        "type" : "object",
        "properties" : {
          "trainNumber" : {
            "type" : "string"
          },
          "startStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "endStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainType" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "pageNum" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ResultListTrainInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TrainInfoDTO"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "TicketInfoDTO" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "trainNumber" : {
            "type" : "string"
          },
          "trainType" : {
            "type" : "string"
          },
          "departureStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "departureStationName" : {
            "type" : "string"
          },
          "arrivalStationId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "arrivalStationName" : {
            "type" : "string"
          },
          "departureDate" : {
            "type" : "string",
            "format" : "date"
          },
          "departureTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "arrivalTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "seatType" : {
            "type" : "string"
          },
          "seatTypeName" : {
            "type" : "string"
          },
          "availableCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "price" : {
            "type" : "number"
          },
          "priceText" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ResultTrainInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TrainInfoDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ResultTicketInfoDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TicketInfoDTO"
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "TicketInventoryQueryDTO" : {
        "type" : "object",
        "properties" : {
          "trainId" : {
            "type" : "integer",
            "description" : "列车ID",
            "format" : "int64",
            "example" : 1
          },
          "seatType" : {
            "type" : "string",
            "description" : "座位类型",
            "example" : "SECOND"
          },
          "status" : {
            "type" : "integer",
            "description" : "状态 0-正常 1-禁用",
            "format" : "int32",
            "example" : 0
          }
        },
        "description" : "票务配置查询条件"
      },
      "ResultListTicketInventoryDTO" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TicketInventoryDTO"
            }
          },
          "timestamp" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}